api-server/pyproject.toml
[tool.poetry]
name = "api-to-do-list"
version = "0.1.0"
description = ""
authors = ["Rhian Pablo <rhianpablo11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
